python-capnp-ffi/requirements.txt
cffi>=1.1.2
pycapnp>=0.5.7
